Added category relation to content.
Content can now also be related to categories. This will become useful
for creating cool archive pages for categories.

+++ Models/Manager/ContentModels/EditModel.cs
@@ -6,6 +6,7 @@ using System.IO;
 using System.Linq;
 using System.Text;
 using System.Web;
+using System.Web.Mvc;
 
 using Piranha.Data;
 
@@ -20,7 +21,17 @@ namespace Piranha.Models.Manager.ContentModels
 		/// <summary>
 		/// Gets/sets the content record.
 		/// </summary>
-		public Piranha.Models.Content Content { get ; set ; }
+		public Models.Content Content { get ; set ; }
+
+		/// <summary>
+		/// Gets/sets the categories associated with the post.
+		/// </summary>
+		public List<Guid> ContentCategories { get ; set ; }
+
+		/// <summary>
+		/// Gets/sets the available categories.
+		/// </summary>
+		public MultiSelectList Categories { get ; set ; }
 
 		/// <summary>
 		/// Gets/sets the optional file.
@@ -33,6 +44,9 @@ namespace Piranha.Models.Manager.ContentModels
 		/// </summary>
 		public EditModel() {
 			Content = new Piranha.Models.Content() ;
+			ContentCategories = new List<Guid>() ;
+			Categories = new MultiSelectList(Category.GetFields("category_id, category_name", 
+				new Params() { OrderBy = "category_name" }), "Id", "Name") ;
 		}
 
 		/// <summary>
@@ -43,6 +57,9 @@ namespace Piranha.Models.Manager.ContentModels
 		public static EditModel GetById(Guid id) {
 			EditModel em = new EditModel() ;
 			em.Content = Piranha.Models.Content.GetSingle(id) ;
+			Relation.GetFieldsByDataId("relation_related_id", id).ForEach(r => em.ContentCategories.Add(r.RelatedId)) ;
+			em.Categories = new MultiSelectList(Category.GetFields("category_id, category_name", 
+				new Params() { OrderBy = "category_name" }), "Id", "Name", em.ContentCategories) ;
 
 			return em ;
 		}
@@ -52,29 +69,39 @@ namespace Piranha.Models.Manager.ContentModels
 		/// </summary>
 		public bool SaveAll() {
 			var context = HttpContext.Current ;
+			var hasfile = UploadedFile != null ;
 
-			// Check if this is an image
-			try {
-				Image img = Image.FromStream(UploadedFile.InputStream) ;
-				Content.IsImage = true ;
-				Content.Width = img.Width ;
-				Content.Height = img.Height ;
-			} catch {
-				Content.IsImage = false ;
+			if (hasfile) {
+				// Check if this is an image
+				try {
+					Image img = Image.FromStream(UploadedFile.InputStream) ;
+					Content.IsImage = true ;
+					Content.Width = img.Width ;
+					Content.Height = img.Height ;
+				} catch {
+					Content.IsImage = false ;
+				}
+				Content.Filename = UploadedFile.FileName ;
+				Content.Type = UploadedFile.ContentType ;
+				Content.Size = UploadedFile.ContentLength ;
 			}
 
-			Content.Filename = UploadedFile.FileName ;
-			Content.Type = UploadedFile.ContentType ;
-			Content.Size = UploadedFile.ContentLength ;
-
 			if (Content.Save()) {
-				string path = context.Server.MapPath("~/App_Data/content") ;
-				if (File.Exists(Content.PhysicalPath)) {
-					File.Delete(Content.PhysicalPath) ;
-					Content.DeleteCache() ;
+				Relation.DeleteByDataId(Content.Id) ;
+				List<Relation> relations = new List<Relation>() ;
+				ContentCategories.ForEach(c => relations.Add(new Relation() { 
+					DataId = Content.Id, RelatedId = c, Type = Relation.RelationType.CONTENTCATEGORY })
+					) ;
+				relations.ForEach(r => r.Save()) ;
+
+				if (hasfile) {
+					string path = context.Server.MapPath("~/App_Data/content") ;
+					if (File.Exists(Content.PhysicalPath)) {
+						File.Delete(Content.PhysicalPath) ;
+						Content.DeleteCache() ;
+					}
+					UploadedFile.SaveAs(Content.PhysicalPath) ;
 				}
-				UploadedFile.SaveAs(Content.PhysicalPath) ;
-				//UploadedFile.SaveAs(path + "/" + Content.Id) ;
 				return true ;
 			}
 			return false ;
@@ -93,6 +120,17 @@ namespace Piranha.Models.Manager.ContentModels
 				} catch { tx.Rollback() ; }
 			}
 			return false ;
+		}
+
+		/// <summary>
+		/// Refreshes the current object.
+		/// </summary>
+		public void Refresh() {
+			if (!Content.IsNew) {
+				Relation.GetFieldsByDataId("relation_related_id", Content.Id).ForEach(r => ContentCategories.Add(r.RelatedId)) ;
+				Categories = new MultiSelectList(Category.GetFields("category_id, category_name", 
+					new Params() { OrderBy = "category_name" }), "Id", "Name", ContentCategories) ;
+			}
 		}
 	}
 }
